refactor phase1: seperate out creation logic into a factory class
