+++ src/Domain.LinnApps/Organisation.cs
@@ -4,18 +4,6 @@
 
     public class Organisation
     {
-        public int OrgId { get; protected set; }
-
-        public string Title { get; protected set; }
-
-        public string VatRegistrationNumber { get; protected set; }
-
-        public string PhoneNumber { get; protected set; }
-
-        public Address Address { get; protected set; } 
-        
-        public DateTime Created { get; protected set; }
-
         public Organisation()
         {
         }
@@ -32,5 +20,17 @@
             this.Address = address;
             this.Created = DateTime.Now;
         }
+
+        public int OrgId { get; protected set; }
+
+        public string Title { get; protected set; }
+
+        public string VatRegistrationNumber { get; protected set; }
+
+        public string PhoneNumber { get; protected set; }
+
+        public Address Address { get; protected set; }
+
+        public DateTime Created { get; protected set; }
     }
 }
