+++ src/Domain.LinnApps/Requisitions/CreationStrategies/ICreationStrategy.cs
@@ -1,0 +1,9 @@
+﻿namespace Linn.Stores2.Domain.LinnApps.Requisitions.CreationStrategies
+{
+    using System.Threading.Tasks;
+
+    public interface ICreationStrategy
+    {
+        Task Apply(RequisitionHeader requisition);
+    }
+}
