+++ src/Domain.LinnApps/Requisitions/IRequisitionFactory.cs
@@ -1,26 +1,9 @@
-namespace Linn.Stores2.Domain.LinnApps.Requisitions
+﻿namespace Linn.Stores2.Domain.LinnApps.Requisitions
 {
     using System.Threading.Tasks;
 
-    public interface IRequisitionService
+    public interface IRequisitionFactory
     {
-        Task<RequisitionHeader> CancelHeader(
-            int reqNumber, 
-            User cancelledBy,
-            string reason,
-            bool requiresAuth = true);
-
-        Task<RequisitionHeader> CancelLine(
-            int reqNumber,
-            int lineNumber,
-            User cancelledBy,
-            string reason);
-
-        Task<RequisitionHeader> BookRequisition(
-            int reqNumber,
-            int? lineNumber,
-            User bookedBy);
-
         Task<RequisitionHeader> CreateRequisition(
             User createdBy,
             string functionCode,
@@ -28,7 +11,7 @@ namespace Linn.Stores2.Domain.LinnApps.Requisitions
             int? document1Number,
             string document1Type,
             string departmentCode,
-            string nominalCode, 
+            string nominalCode,
             LineCandidate firstLine = null,
             string reference = null,
             string comments = null,
