+++ tests/Integration/Integration.Tests/RequisitionModuleTests/WhenCancellingLine.cs
@@ -39,7 +39,7 @@
                 new Department(),
                 new Nominal());
             req.Cancel("reasonable reason", new Employee());
-            this.DomainService.CancelLine(
+            this.ReqManager.CancelLine(
                     this.resource.ReqNumber, this.resource.LineNumber.Value, Arg.Any<User>(), this.resource.Reason)
                 .Returns(req);
             this.Response = this.Client.PostAsJsonAsync("/requisitions/cancel", this.resource).Result;
@@ -48,7 +48,7 @@
         [Test]
         public void ShouldCancelLine()
         {
-            this.DomainService.Received(1).CancelLine(
+            this.ReqManager.Received(1).CancelLine(
                 this.resource.ReqNumber, 
                 this.resource.LineNumber.Value, 
                 Arg.Any<User>(),
