+++ tests/Unit/Domain.LinnApps.Tests/RequisitionFactoryTests/WhenCreatingLdReqAndCreateNominalPostingFails.cs
@@ -1,4 +1,4 @@
-﻿namespace Linn.Stores2.Domain.LinnApps.Tests.RequisitionServiceTests
+﻿namespace Linn.Stores2.Domain.LinnApps.Tests.RequisitionFactoryTests
 {
     using System;
     using System.Collections.Generic;
@@ -7,7 +7,6 @@
 
     using FluentAssertions;
 
-    using Linn.Common.Domain;
     using Linn.Stores2.Domain.LinnApps.Accounts;
     using Linn.Stores2.Domain.LinnApps.Exceptions;
     using Linn.Stores2.Domain.LinnApps.Parts;
@@ -16,10 +15,11 @@
     using Linn.Stores2.TestData.Requisitions;
 
     using NSubstitute;
+    using NSubstitute.ExceptionExtensions;
 
     using NUnit.Framework;
 
-    public class WhenFailToCreateAndFailToCancelHeader : ContextBase
+    public class WhenCreatingLdreqFromAndCreateNominalPostingStockFails : ContextBase
     {
         private User user;
         private StoresFunction ldreq;
@@ -61,8 +61,6 @@
 
             this.NominalRepository.FindByIdAsync(this.nominal.NominalCode).Returns(this.nominal);
 
-            this.TransactionDefinitionRepository.FindByIdAsync("TRANS").Returns(this.transactionDefinition);
-
             this.StorageLocationRepository.FindByAsync(Arg.Any<Expression<Func<StorageLocation, bool>>>())
                 .Returns(this.from);
 
@@ -78,26 +76,10 @@
                 this.department,
                 this.nominal);
 
-            this.ReqStoredProcedures
-                .PickStock(
-                    this.part.PartNumber,
-                    Arg.Any<int>(),
-                    1,
-                    1,
-                    this.from.LocationId,
-                    null,
-                    "LINN",
-                    this.transactionDefinition.TransactionCode)
-                .Returns(new ProcessResult(true, string.Empty));
-
-            this.ReqStoredProcedures.CreateNominals(
-                    Arg.Any<int>(), 1, 1, this.nominal.NominalCode, this.department.DepartmentCode)
-                .Returns(new ProcessResult(false, "can't post there"));
-
-            this.ReqStoredProcedures.DeleteAllocOntos(Arg.Any<int>(), null, Arg.Any<int>(), this.createdReq.Document1Name)
-                .Returns(new ProcessResult(false, "couldn't un-allocate stock!"));
-
             this.ReqRepository.FindByIdAsync(Arg.Any<int>()).Returns(this.createdReq);
+
+            this.RequisitionManager.AddRequisitionLine(Arg.Any<RequisitionHeader>(), Arg.Any<LineCandidate>())
+                .Throws(new CreateNominalPostingException("failed in create_nominals - cant post this to that"));
 
             this.action = async () => await this.Sut.CreateRequisition(
                 this.user,
@@ -119,7 +101,7 @@
                     LineNumber = 1,
                     PartNumber = this.part.PartNumber,
                     Qty = 1,
-                    TransactionDefinition = transactionDefinition.TransactionCode
+                    TransactionDefinition = this.transactionDefinition.TransactionCode
                 },
                 "ref",
                 "comments",
@@ -133,8 +115,18 @@
             await this.action.Should()
                 .ThrowAsync<CreateRequisitionException>()
                 .WithMessage(
-                    "Warning - req failed to create: failed in create_nominals: can't post there. "
-                    + "Header also failed to cancel: couldn't un-allocate stock!. Some cleanup may be required!");
+                    "Req failed to create since first line could not be added. Reason: failed in create_nominals - cant post this to that");
+        }
+
+        [Test]
+        public void ShouldCancelHeader()
+        {
+            this.action();
+            this.RequisitionManager.Received(1).CancelHeader(
+                Arg.Any<int>(),
+                Arg.Any<User>(),
+                "Req failed to create since first line could not be added. Reason: failed in create_nominals - cant post this to that",
+                false);
         }
     }
 }
