+++ tests/Unit/Domain.LinnApps.Tests/RequisitionManagerTests/WhenBooking.cs
@@ -1,13 +1,16 @@
-﻿namespace Linn.Stores2.Domain.LinnApps.Tests.RequisitionServiceTests
+﻿namespace Linn.Stores2.Domain.LinnApps.Tests.RequisitionManagerTests
 {
     using System.Collections.Generic;
 
     using FluentAssertions;
+
     using Linn.Common.Domain;
     using Linn.Stores2.Domain.LinnApps.Accounts;
     using Linn.Stores2.Domain.LinnApps.Requisitions;
     using Linn.Stores2.TestData.Requisitions;
+
     using NSubstitute;
+
     using NUnit.Framework;
 
     public class WhenBooking : ContextBase
