+++ tests/Unit/Domain.LinnApps.Tests/RequisitionManagerTests/WhenBookingAndDoRequisitionFails.cs
@@ -1,16 +1,19 @@
-﻿namespace Linn.Stores2.Domain.LinnApps.Tests.RequisitionServiceTests
+﻿namespace Linn.Stores2.Domain.LinnApps.Tests.RequisitionManagerTests
 {
     using System;
     using System.Collections.Generic;
     using System.Threading.Tasks;
+
     using FluentAssertions;
 
+    using Linn.Common.Domain;
     using Linn.Stores2.Domain.LinnApps.Exceptions;
 
     using NSubstitute;
+
     using NUnit.Framework;
 
-    public class WhenBookingAndUnauthorised : ContextBase
+    public class WhenBookingAndDoRequisitionFails : ContextBase
     {
         private Func<Task> action;
 
@@ -25,14 +28,20 @@
 
             this.AuthService.HasPermissionFor(
                     AuthorisedActions.BookRequisition, Arg.Any<IEnumerable<string>>())
-                .Returns(false);
+                .Returns(true);
+
+            this.ReqStoredProcedures.DoRequisition(
+                123,
+                null,
+                user.UserNumber).Returns(new ProcessResult(false, "Stores failure"));
+
             this.action = async () => await this.Sut.BookRequisition(123, null, user);
         }
 
         [Test]
         public async Task ShouldThrow()
         {
-            await this.action.Should().ThrowAsync<UnauthorisedActionException>();
+            await this.action.Should().ThrowAsync<RequisitionException>();
         }
     }
 }
