+++ tests/Unit/Domain.LinnApps.Tests/RequisitionManagerTests/WhenCancelling.cs
@@ -1,4 +1,4 @@
-namespace Linn.Stores2.Domain.LinnApps.Tests.RequisitionServiceTests
+namespace Linn.Stores2.Domain.LinnApps.Tests.RequisitionManagerTests
 {
     using System.Collections.Generic;
 
@@ -7,13 +7,14 @@ namespace Linn.Stores2.Domain.LinnApps.Tests.RequisitionServiceTests
     using Linn.Common.Domain;
     using Linn.Stores2.Domain.LinnApps.Accounts;
     using Linn.Stores2.Domain.LinnApps.Requisitions;
+    using Linn.Stores2.Domain.LinnApps.Tests.RequisitionManagerTests;
     using Linn.Stores2.TestData.Requisitions;
 
     using NSubstitute;
 
     using NUnit.Framework;
 
-    public class WhenCancellingLine : ContextBase
+    public class WhenCancelling : ContextBase
     {
         private RequisitionHeader req;
 
@@ -23,17 +24,13 @@ namespace Linn.Stores2.Domain.LinnApps.Tests.RequisitionServiceTests
             this.req = new ReqWithReqNumber(
                 123,
                 new Employee(),
-                new StoresFunction { FunctionCode = "F1" },
+                new StoresFunction { FunctionCode = "FUNC" },
                 "F",
-                12345678,
-                "TYPE",
-                new Department(),
-                new Nominal(),
-                null,
-                null,
-                "Goodbye Reqs");
-            var requisitionLine = new RequisitionLine(this.req.ReqNumber, 1);
-            this.req.AddLine(requisitionLine);
+                123,
+                "REQ",
+                new Department(), 
+                new Nominal());
+
             this.ReqRepository.FindByIdAsync(this.req.ReqNumber).Returns(this.req);
 
             var user = new User
@@ -45,15 +42,14 @@ namespace Linn.Stores2.Domain.LinnApps.Tests.RequisitionServiceTests
 
             this.ReqStoredProcedures.DeleteAllocOntos(
                 this.req.ReqNumber,
-                1,
+                null,
                 this.req.Document1.GetValueOrDefault(),
                 this.req.Document1Name).Returns(new ProcessResult(true, string.Empty));
 
             this.AuthService.HasPermissionFor(
                 AuthorisedActions.CancelRequisition, Arg.Any<IEnumerable<string>>()).Returns(true);
-            this.req = this.Sut.CancelLine(
-                123, 
-                1,
+            this.req = this.Sut.CancelHeader(
+                this.req.ReqNumber, 
                 user,
                 "REASON").Result;
         }
@@ -64,7 +60,7 @@ namespace Linn.Stores2.Domain.LinnApps.Tests.RequisitionServiceTests
             this.ReqStoredProcedures.Received()
                 .DeleteAllocOntos(
                     this.req.ReqNumber,
-                    1,
+                    null,
                     this.req.Document1.GetValueOrDefault(),
                     this.req.Document1Name);
         }
@@ -72,6 +68,7 @@ namespace Linn.Stores2.Domain.LinnApps.Tests.RequisitionServiceTests
         [Test]
         public void ShouldReturnCancelled()
         {
+            this.req.ReqNumber.Should().Be(123);
             this.req.Cancelled.Should().Be("Y");
             this.req.CancelDetails.Count.Should().Be(1);
         }
