+++ tests/Unit/Domain.LinnApps.Tests/RequisitionManagerTests/WhenCancellingAndUnauthorised.cs
@@ -1,4 +1,4 @@
-namespace Linn.Stores2.Domain.LinnApps.Tests.RequisitionServiceTests
+namespace Linn.Stores2.Domain.LinnApps.Tests.RequisitionManagerTests
 {
     using System;
     using System.Collections.Generic;
@@ -7,12 +7,13 @@ namespace Linn.Stores2.Domain.LinnApps.Tests.RequisitionServiceTests
     using FluentAssertions;
 
     using Linn.Stores2.Domain.LinnApps.Exceptions;
+    using Linn.Stores2.Domain.LinnApps.Tests.RequisitionManagerTests;
 
     using NSubstitute;
 
     using NUnit.Framework;
 
-    public class WhenCancellingLineAndUnauthorised : ContextBase
+    public class WhenCancellingAndUnauthorised : ContextBase
     {
         private Func<Task> action;
         
@@ -28,7 +29,7 @@ namespace Linn.Stores2.Domain.LinnApps.Tests.RequisitionServiceTests
             this.AuthService.HasPermissionFor(
                 AuthorisedActions.CancelRequisition, Arg.Any<IEnumerable<string>>())
                 .Returns(false);
-            this.action = async () => await this.Sut.CancelLine(123, 1, user, "REASON");
+            this.action = async () => await this.Sut.CancelHeader(123, user, "REASON");
         }
 
         [Test]
