+++ tests/Unit/Domain.LinnApps.Tests/RequisitionManagerTests/WhenCancellingLineAndUnallocatingFails.cs
@@ -1,35 +1,42 @@
-namespace Linn.Stores2.Domain.LinnApps.Tests.RequisitionServiceTests
+namespace Linn.Stores2.Domain.LinnApps.Tests.RequisitionManagerTests
 {
+    using System;
     using System.Collections.Generic;
+    using System.Threading.Tasks;
 
     using FluentAssertions;
 
     using Linn.Common.Domain;
     using Linn.Stores2.Domain.LinnApps.Accounts;
+    using Linn.Stores2.Domain.LinnApps.Exceptions;
     using Linn.Stores2.Domain.LinnApps.Requisitions;
+    using Linn.Stores2.Domain.LinnApps.Tests.RequisitionManagerTests;
     using Linn.Stores2.TestData.Requisitions;
 
     using NSubstitute;
 
     using NUnit.Framework;
 
-    public class WhenCancelling : ContextBase
+    public class WhenCancellingLineAndUnallocatingFails : ContextBase
     {
         private RequisitionHeader req;
+        
+        private Func<Task> action;
 
         [SetUp]
         public void SetUp()
         {
             this.req = new ReqWithReqNumber(
-                123,
-                new Employee(),
-                new StoresFunction { FunctionCode = "FUNC" },
-                "F",
-                123,
-                "REQ",
-                new Department(), 
-                new Nominal());
-
+                    123,
+                    new Employee(),
+                    new StoresFunction { FunctionCode = "FUNC" },
+                    "F",
+                    123,
+                    "REQ",
+                    new Department(),
+                    new Nominal());
+            var requisitionLine = new RequisitionLine(this.req.ReqNumber, 1);
+            this.req.AddLine(requisitionLine);
             this.ReqRepository.FindByIdAsync(this.req.ReqNumber).Returns(this.req);
 
             var user = new User
@@ -41,35 +48,20 @@ namespace Linn.Stores2.Domain.LinnApps.Tests.RequisitionServiceTests
 
             this.ReqStoredProcedures.DeleteAllocOntos(
                 this.req.ReqNumber,
-                null,
+                1,
                 this.req.Document1.GetValueOrDefault(),
-                this.req.Document1Name).Returns(new ProcessResult(true, string.Empty));
+                this.req.Document1Name).Returns(new ProcessResult(false, "Some stores-y message"));
 
             this.AuthService.HasPermissionFor(
                 AuthorisedActions.CancelRequisition, Arg.Any<IEnumerable<string>>()).Returns(true);
-            this.req = this.Sut.CancelHeader(
-                this.req.ReqNumber, 
-                user,
-                "REASON").Result;
+            this.action = async () => await this.Sut.CancelLine(123, 1, user, "REASON");
         }
 
         [Test]
-        public void ShouldUnallocateStock()
+        public async Task ShouldThrowError()
         {
-            this.ReqStoredProcedures.Received()
-                .DeleteAllocOntos(
-                    this.req.ReqNumber,
-                    null,
-                    this.req.Document1.GetValueOrDefault(),
-                    this.req.Document1Name);
-        }
-
-        [Test]
-        public void ShouldReturnCancelled()
-        {
-            this.req.ReqNumber.Should().Be(123);
-            this.req.Cancelled.Should().Be("Y");
-            this.req.CancelDetails.Count.Should().Be(1);
+            await this.action.Should().ThrowAsync<RequisitionException>()
+                .WithMessage("Some stores-y message");
         }
     }
 }
